리뷰 별점 수정

+++ Uglory_Customer/src/main/java/edu/kh/ugloryC/review/model/vo/ReviewSelectInfo.java
@@ -21,7 +21,7 @@ public class ReviewSelectInfo {
 	private int reviewNo;
 	private String reviewContent;
 	private String reviewEnrollDate;
-	private int starRating;
+	private double starRating;
 	
 	private int memberNo;
 	private String memberName;
